refactor: split github-checkout.ts into smaller modules
- Extract GitHub API calls to github-api.ts
- Create separate modules for PR and Issue checkout logic
- Add tests for each module
- Improve code organization and maintainability
