+++ packages/github/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./github-checkout.ts";
+export type { GitHubPullRequest, GitHubIssue } from "./github-api.ts";
